:truck: move tests to separate files
